+++ src/invoices/InvoiceDetails/VoucherInformation/VoucherInformation.js
@@ -20,7 +20,7 @@ const columnMapping = {
   amount: <FormattedMessage id="ui-invoice.invoice.details.voucher.amount" />,
 };
 
-const VoucherInformation = ({ voucher, voucherLines = [] }) => {
+const VoucherInformation = ({ voucher, voucherLines }) => {
   const groupedVoucherLines = groupByExternalAccNumber(voucherLines);
   const linesWithTotalAmount = Object.values(groupedVoucherLines).map(lines => ({
     totalAmount: getTotalAmount(lines),
@@ -60,6 +60,10 @@ const VoucherInformation = ({ voucher, voucherLines = [] }) => {
 VoucherInformation.propTypes = {
   voucher: PropTypes.object.isRequired,
   voucherLines: PropTypes.arrayOf(PropTypes.object),
+};
+
+VoucherInformation.defaultProps = {
+  voucherLines: [],
 };
 
 export default VoucherInformation;